AutoCoder/requirements.txt
absl-py==0.10.0
appnope==0.1.0
astunparse==1.6.3
attrs==19.3.0
backcall==0.1.0
bleach==3.1.5
blis==0.4.1
bokeh==2.2.1
boto==2.49.0
boto3==1.14.56
botocore==1.17.56
cachetools==4.1.1
catalogue==1.0.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cycler==0.10.0
cymem==2.0.3
decorator==4.4.2
defusedxml==0.6.0
distro==1.5.0
docutils==0.15.2
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.3.1/en_core_web_sm-2.3.1.tar.gz
entrypoints==0.3
future==0.18.2
fuzzywuzzy==0.18.0
gast==0.3.3
gensim==3.8.3
google-auth==1.21.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.31.0
h5py==2.10.0
idna==2.9
ipykernel==5.2.1
ipython==7.14.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.0
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.16.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.3
jupyter-console==6.1.0
jupyter-core==4.6.3
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.3.1
mistune==0.8.4
murmurhash==1.0.2
nbconvert==5.6.1
nbformat==5.0.6
networkx==2.5
nltk==3.5
notebook==6.0.3
numpy==1.18.4
oauthlib==3.1.0
opt-einsum==3.3.0
packaging==20.3
pandas==1.0.3
pandocfilters==1.4.2
parso==0.7.0
pdfminer==20191125
pdfminer.six==20200402
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
pke @ git+https://github.com/boudinfl/pke.git@ae0d7567094ea3729567d5f1af63bec0b138e4d5
plac==1.1.3
preshed==3.0.2
prometheus-client==0.7.1
prompt-toolkit==3.0.5
protobuf==3.13.0
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.9.7
Pygments==2.6.1
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
pyzmq==19.0.1
qtconsole==4.7.3
QtPy==1.9.0
regex==2020.7.14
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.6
s3transfer==0.3.3
scikit-learn==0.23.2
scipy==1.4.1
seaborn==0.10.1
Send2Trash==1.5.0
simpleneighbors==0.1.0
six==1.14.0
sklearn==0.0
smart-open==2.1.1
sortedcontainers==2.1.0
spacy==2.3.2
srsly==1.0.2
tabula-py==2.1.1
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.0
tensorflow-estimator==2.3.0
tensorflow-hub==0.9.0
tensorflow-text==2.3.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
textblob==0.15.3
tf-sentencepiece==0.1.90
thinc==7.4.1
threadpoolctl==2.1.0
tornado==6.0.4
tqdm==4.48.2
traitlets==4.3.3
typing-extensions==3.7.4.3
Unidecode==1.1.1
urllib3==1.25.9
wasabi==0.7.1
wcwidth==0.1.9
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wordcloud==1.8.0
wrapt==1.12.1
